+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [0.122.0](https://github.com/go-to-k/cdkd/compare/v0.121.0...v0.122.0) (2026-05-22)
+
+
+### Features
+
+* **local:** --from-state resolves Fn::ImportValue and Fn::GetStackOutput ([#454](https://github.com/go-to-k/cdkd/issues/454)) ([#487](https://github.com/go-to-k/cdkd/issues/487)) ([628d5e8](https://github.com/go-to-k/cdkd/commit/628d5e87bb7490f1960a9db7109d7cb64fd7143d))
+
+# [0.121.0](https://github.com/go-to-k/cdkd/compare/v0.120.0...v0.121.0) (2026-05-22)
+
+
+### Features
+
+* **logger:** rich color output + replace hero GIF with side-by-side demo ([#486](https://github.com/go-to-k/cdkd/issues/486)) ([9dee8dc](https://github.com/go-to-k/cdkd/commit/9dee8dcddcd5b7075309005e923dc517b8b1557f))
+
+# [0.120.0](https://github.com/go-to-k/cdkd/compare/v0.119.0...v0.120.0) (2026-05-22)
+
+
+### Features
+
+* **local-start-api:** REST v1 IAM authorizer SigV4 signature verification ([#447](https://github.com/go-to-k/cdkd/issues/447)) ([#484](https://github.com/go-to-k/cdkd/issues/484)) ([b1a379c](https://github.com/go-to-k/cdkd/commit/b1a379c107b6797e6903c4fee67655300b177884))
+
+# [0.119.0](https://github.com/go-to-k/cdkd/compare/v0.118.0...v0.119.0) (2026-05-22)
+
+
+### Features
+
+* **local-start-api:** cognito multi-pool federation ([#488](https://github.com/go-to-k/cdkd/issues/488)) ([64a479a](https://github.com/go-to-k/cdkd/commit/64a479a91a7d0af3fdf33fa924d8029196035f34))
+
 # [0.118.0](https://github.com/go-to-k/cdkd/compare/v0.117.2...v0.118.0) (2026-05-21)